dll_98_func03 99.615->99.769: spell the frameValues[2] mirror with an explicit (s32) promotion -- the plain re-read of the just-stored s16 field is store-forwarded (no instruction) where retail re-materializes the canonical s16 value in the slot; the cast gives the read a distinct value number so an extsh lands exactly where retail reloads (lha). True retail shape is a nocse reload (plain re-read under opt_common_subs off) -- TU re-split candidate for the pragma-carve lane; this is the best source-reachable form under cflags_dll_noopt.

+++ src/main/dll/dll_0098_dll98func0.c
@@ -45,7 +45,7 @@ void dll_98_func03(int sourceObj, int variant, int posSource, u32 flags, int arg
     GfxCmd* entry;
     int anim;
     effectTable->frameValues[1] = randomGetRange(0, 0x1e) + 0x1e;
-    effectTable->frameValues[2] = effectTable->frameValues[1];
+    effectTable->frameValues[2] = (s32)effectTable->frameValues[1];
     entry = buf.entries;
     entry[0].layer = 0;
     entry[0].flags = 0x12;
